# server/admin/docs/swagger.yaml
basePath: /api/v1/admin
definitions:
  admin.ErrorResponse:
    properties:
      error:
        example: resource not found
        type: string
    type: object
...
  store.BackendRef:
    properties:
      backend:
        type: string
      model:
        type: string
    type: object
...
  store.MemoryProvider:
    properties:
      category:
        type: string
      config:
        additionalProperties: true
        type: object
      embedding:
        $ref: '#/definitions/store.BackendRef'
      name:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Administration API for managing Magec multi-agent voice assistant resources.
  title: Magec Admin API
  version: "1.0"
paths:
...
  /memory:
    get:
      description: Returns all configured memory providers (session and long-term)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.MemoryProvider'
            type: array
      summary: List memory providers
      tags:
      - memory
    post:
      consumes:
      - application/json
      description: Creates a new memory provider. Type must be registered and support
        the given category.
      parameters:
      - description: Memory provider definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.MemoryProvider'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.MemoryProvider'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Create memory provider
      tags:
      - memory
  /memory/{name}:
    delete:
      description: Deletes a memory provider by name
      parameters:
      - description: Provider name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Delete memory provider
      tags:
      - memory
    get:
      description: Returns a memory provider by its unique name
      parameters:
      - description: Provider name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.MemoryProvider'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Get memory provider
      tags:
      - memory
    put:
      consumes:
      - application/json
      description: Updates a memory provider by name
      parameters:
      - description: Provider name
        in: path
        name: name
        required: true
        type: string
      - description: Memory provider definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/store.MemoryProvider'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.MemoryProvider'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.ErrorResponse'
      summary: Update memory provider
      tags:
      - memory
...
schemes:
- http
swagger: "2.0"